backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
Pillow==11.2.1
django-cors-headers==4.5.0
psycopg2-binary==2.9.3